block filter, keep filter simple, block templatetag
the {% filter fn( FILTER_STR ) %} ... {% end %} evaluates the block and
outputs in the FILTER_STR variable, with this the user can apply function
(filters) on it.

keep filters simple: filters are just lua functions loaded in the
Context that can be applied inside {{ ... }} or in the FILTER_STR of a
{% filter %} tag
  + a file that returns a table of filter can be loaded by the context
with add_filter( 'path.to.filter' ) and can be accessed by the context
'filter'. letk.template.filters is always loaded.

{% templatetag <name> %} inserts a template tag as {%, %}, {{, }}, {#,
etc